enhancement(hardhatconfig): break apart into seperate files
