+++ contracts/dollar/deploy/CurveUADIncentive.ts
@@ -1,10 +1,10 @@
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { DeployFunction } from "hardhat-deploy/types";
-
+import { warn } from "../hardhat.config";
 const func: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
   const { deployments, ethers } = hre;
   const [admin] = await ethers.getSigners();
-  deployments.log("admin address :", admin.address);
+  deployments.log(warn(`admin.address: ${admin.address}`));
 
   const opts = {
     from: admin.address,
