+++ contracts/dollar/deploy/DebtCoupon.ts
@@ -2,11 +2,11 @@ import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { DeployFunction } from "hardhat-deploy/types";
 import { UbiquityAlgorithmicDollarManager } from "../artifacts/types/UbiquityAlgorithmicDollarManager";
 import pressAnyKey from "../utils/flow";
-
+import { warn } from "../hardhat.config";
 const func: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
   const { deployments, ethers } = hre;
   const [admin] = await ethers.getSigners();
-  deployments.log("admin address :", admin.address);
+  deployments.log(warn(`admin.address: ${admin.address}`));
 
   const opts = {
     from: admin.address,
