+++ contracts/dollar/hardhat-config/utils/warn.ts
@@ -1,0 +1,7 @@
+import { colorizeText } from "../tasks/utils/console-colors";
+
+
+
+export function warn(message: string) {
+    console.warn(colorizeText(`\t⚠ ${message}`, "fgYellow"));
+}
